deployed-django-projects/Portfolio/pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = ""
authors = ["zoelinsg <zoelin.sg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4"
django = "^5.1.5"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-environ = {version = "^0.12.0", python = ">=3.9,<4"}
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pillow = "^10.0.0"
djangorestframework-simplejwt = "^5.4.0"
django-redis = "^5.4.0"
django-filter = "^24.3"
django-taggit = "^6.1.0"
django-markdownx = "^4.0.7"
moviepy = "^2.1.2"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
